Fix broken `uint8_t*` interface that truncates at the first NULL byte
The generated Cython code truncates the key at the first NULL, parsing
it as a null terminated string.

We still offer a similar interface, the idiomatic one that operates on
a copy of the input data, calling `bytes(data)`, and a fast one that
operates on `bytearray`s only, mutating the data inplace.